requirements.txt
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.20.1  --index-url https://download.pytorch.org/whl/cu118
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sympy==1.13.1
torchinfo==1.8.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
zipp==3.23.0
